requirements.txt
appdirs==1.4.3
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
black==19.10b0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dialogflow==0.5.2
Flask==1.0.2
google-api-core==1.7.0
google-auth==1.6.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==6.0.0
oauthlib==3.0.1
pathspec==0.6.0
pluggy==0.8.1
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.2.2
pytest==4.3.0
python-dotenv==0.10.1
pytz==2018.9
regex==2019.12.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
